Refactor: Simplify chat settings context
