+++ src/context/index.ts
@@ -1,3 +1,3 @@
 export { type AvatarContextType } from './avatar-context.tsx';
-export { type ChatSettingsContextType } from './chat-settings-context.tsx';
+export { type ChatSettings, type ChatSettingsContextType } from './chat-settings-context.tsx';
 export { type ThemeContextType } from './theme-context.tsx';
